services/user_and_teams_service/pyproject.toml
[tool.poetry]
name = "hackathons-platform"
version = "0.1.0"
description = ""
authors = ["Maxim Drugov <midrugov@edu.hse.ru>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10"
fastapi = {extras = ["all"], version = "^0.115.5"}
uvicorn = {extras = ["standard"], version = "^0.31.0"}
pydantic = {extras = ["email"], version = "^2.9.2"}
ruff = "^0.6.9"
coverage = "^7.6.3"
sqlalchemy = "^2.0.35"
alembic = "^1.13.3"
pydantic-settings = "^2.6.0"
pyjwt = "^2.9.0"
passlib = "^1.7.4"
pre-commit = "^4.0.1"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.0.1"
asyncpg = "^0.30.0"
aiosqlite = "^0.20.0"
pytest-asyncio = "^0.24.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
ruff = "^0.6.9"
black = "^24.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py310"
exclude = [
    "services/user_and_teams_service/app/alembic/versions",
]
